@@ -90,7 +90,7 @@ namespace TestCLI
                         height = size;
                     }
 
-                    using (var resized = original.Resize(new SKImageInfo(width, height), SKBitmapResizeMethod.Lanczos3))
+                    using (var resized = original.Resize(new SKImageInfo(width, height), SKFilterQuality.Medium))
                     {
                         // No need to resize
                         if (resized == null)
