@@ -1,96 +1,26 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 //
-// Generated with Bot Builder V4 SDK Template for Visual Studio EchoBot v4.5.0
+// Generated with Bot Builder V4 SDK Template for Visual Studio EchoBot v4.10.3
 
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Bot.Builder;
-using Microsoft.Bot.Schema;
 using Microsoft.Bot.Builder.Dialogs;
-using Microsoft.Extensions.Logging;
-using System.Linq;
-using PictureBot.Models;
-using PictureBot.Responses;
-using Microsoft.Bot.Builder.AI.Luis;
-using Microsoft.Azure.Search;
-using Microsoft.Azure.Search.Models;
-using System;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
+using Microsoft.Bot.Schema;
 using Microsoft.PictureBot;
+using PictureBot.Responses;
 
 namespace PictureBot.Bots
 {
-
-    /// <summary>
-    /// Represents a bot that processes incoming activities.
-    /// For each user interaction, an instance of this class is created and the OnTurnAsync method is called.
-    /// This is a Transient lifetime service.  Transient lifetime services are created
-    /// each time they're requested. For each Activity received, a new instance of this
-    /// class is created. Objects that are expensive to construct, or have a lifetime
-    /// beyond the single turn, should be carefully managed.
-    /// For example, the <see cref="MemoryStorage"/> object and associated
-    /// <see cref="IStatePropertyAccessor{T}"/> object are created with a singleton lifetime.
-    /// </summary>
-    /// <seealso cref="https://docs.microsoft.com/en-us/aspnet/core/fundamentals/dependency-injection?view=aspnetcore-2.1"/>
-    /// <summary>Contains the set of dialogs and prompts for the picture bot.</summary>
     public class PictureBot : ActivityHandler
     {
         private readonly PictureBotAccessors _accessors;
-        // Initialize LUIS Recognizer
-
-        private readonly ILogger _logger;
         private DialogSet _dialogs;
 
-        /// <summary>
-        /// Every conversation turn for our PictureBot will call this method.
-        /// There are no dialogs used, since it's "single turn" processing, meaning a single
-        /// request and response. Later, when we add Dialogs, we'll have to navigate through this method.
-        /// </summary>
-        /// <param name="turnContext">A <see cref="ITurnContext"/> containing all the data needed
-        /// for processing this conversation turn. </param>
-        /// <param name="cancellationToken">(Optional) A <see cref="CancellationToken"/> that can be used by other objects
-        /// or threads to receive notice of cancellation.</param>
-        /// <returns>A <see cref="Task"/> that represents the work queued to execute.</returns>
-        /// <seealso cref="BotStateSet"/>
-        /// <seealso cref="ConversationState"/>
-        /// <seealso cref="IMiddleware"/>
-        public override async Task OnTurnAsync(ITurnContext turnContext, CancellationToken cancellationToken = default(CancellationToken))
+        public PictureBot(PictureBotAccessors accessors)
         {
-            if (turnContext.Activity.Type is "message")
-            {
-                // Establish dialog context from the conversation state.
-                var dc = await _dialogs.CreateContextAsync(turnContext);
-                // Continue any current dialog.
-                var results = await dc.ContinueDialogAsync(cancellationToken);
-
-                // Every turn sends a response, so if no response was sent,
-                // then there no dialog is currently active.
-                if (!turnContext.Responded)
-                {
-                    // Start the main dialog
-                    await dc.BeginDialogAsync("mainDialog", null, cancellationToken);
-                }
-            }
-        }
-        /// <summary>
-        /// Initializes a new instance of the <see cref="PictureBot"/> class.
-        /// </summary>
-        /// <param name="accessors">A class containing <see cref="IStatePropertyAccessor{T}"/> used to manage state.</param>
-        /// <param name="loggerFactory">A <see cref="ILoggerFactory"/> that is hooked to the Azure App Service provider.</param>
-        /// <seealso cref="https://docs.microsoft.com/en-us/aspnet/core/fundamentals/logging/?view=aspnetcore-2.1#windows-eventlog-provider"/>
-        public PictureBot(PictureBotAccessors accessors, ILoggerFactory loggerFactory /*, LuisRecognizer recognizer*/)
-        {
-            if (loggerFactory == null)
-            {
-                throw new System.ArgumentNullException(nameof(loggerFactory));
-            }
-
-            // Add instance of LUIS Recognizer
-
-            _logger = loggerFactory.CreateLogger<PictureBot>();
-            _logger.LogTrace("PictureBot turn start.");
             _accessors = accessors ?? throw new System.ArgumentNullException(nameof(accessors));
 
             // The DialogSet needs a DialogState accessor, it will call it when it has a turn context.
@@ -119,14 +49,38 @@ namespace PictureBot.Bots
             _dialogs.Add(new TextPrompt("searchPrompt"));
         }
 
-        // Add MainDialog-related tasks
 
-        // Add SearchDialog-related tasks
+        public override async Task OnTurnAsync(ITurnContext turnContext, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            if (turnContext.Activity.Type is "message")
+            {
+                // Establish dialog context from the conversation state.
+                var dc = await _dialogs.CreateContextAsync(turnContext);
+                // Continue any current dialog.
+                var results = await dc.ContinueDialogAsync(cancellationToken);
 
-        // Add search related tasks
+                // Every turn sends a response, so if no response was sent,
+                // then there no dialog is currently active.
+                if (!turnContext.Responded)
+                {
+                    // Start the main dialog
+                    await dc.BeginDialogAsync("mainDialog", null, cancellationToken);
+                }
+            }
+        }
 
-        // If we haven't greeted a user yet, we want to do that first, but for the rest of the
-        // conversation we want to remember that we've already greeted them.
+        protected override async Task OnMembersAddedAsync(IList<ChannelAccount> membersAdded, ITurnContext<IConversationUpdateActivity> turnContext, CancellationToken cancellationToken)
+        {
+            var welcomeText = "Hello and welcome!";
+            foreach (var member in membersAdded)
+            {
+                if (member.Id != turnContext.Activity.Recipient.Id)
+                {
+                    await turnContext.SendActivityAsync(MessageFactory.Text(welcomeText, welcomeText), cancellationToken);
+                }
+            }
+        }
+
         private async Task<DialogTurnResult> GreetingAsync(WaterfallStepContext stepContext, CancellationToken cancellationToken)
         {
             // Get the state for the current step in the conversation
