@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 //
-// Generated with Bot Builder V4 SDK Template for Visual Studio EchoBot v4.5.0
+// Generated with Bot Builder V4 SDK Template for Visual Studio EchoBot v4.10.3
 
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
@@ -26,7 +26,7 @@ namespace PictureBot.Controllers
             Bot = bot;
         }
 
-        [HttpPost]
+        [HttpPost, HttpGet]
         public async Task PostAsync()
         {
             // Delegate the processing of the HTTP POST to the adapter.
