@@ -1,8 +1,7 @@
-﻿using Microsoft.Azure.Search.Models;
-
+﻿
 namespace PictureBot.Models
 {
-    public class ImageMapper 
+    public class ImageMapper
     {
         public static SearchHit ToSearchHit(dynamic hit)
         {
