@@ -1,5 +1,4 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
+﻿using System.Collections.Generic;
 
 namespace Microsoft.PictureBot
 {
