@@ -7,7 +7,7 @@ namespace PictureBot.Responses
     {
         public static async Task ReplyWithGreeting(ITurnContext context)
         {
-            // Add a greeting
+            await context.SendActivityAsync("Hello, Im a Picture Bot");
         }
         public static async Task ReplyWithHelp(ITurnContext context)
         {
@@ -21,7 +21,7 @@ namespace PictureBot.Responses
         {
             // Add a response for the user if Regex or LUIS doesn't know
             // What the user is trying to communicate
-            await context.SendActivityAsync($"I'm not understand you.");
+            await context.SendActivityAsync($"I'm sorry, I don't understand.");
         }
         public static async Task ReplyWithLuisScore(ITurnContext context, string key, double score)
         {
@@ -31,14 +31,13 @@ namespace PictureBot.Responses
         {
             await context.SendActivityAsync($"Posting your picture(s) on twitter...");
         }
-        public static async Task ReplyWithSearchConfirmation(ITurnContext context)
-        {
-            await context.SendActivityAsync($"Searching your picture(s) on twitter...");
-        }
-        
         public static async Task ReplyWithOrderConfirmation(ITurnContext context)
         {
             await context.SendActivityAsync($"Ordering standard prints of your picture(s)...");
+        }
+        public static async Task ReplyWithSearchConfirmation(ITurnContext context)
+        {
+            await context.SendActivityAsync($"Searching picture(s)...");
         }
     }
 }
